Update Junkyard Events.i7x
Updated Stripped Car event so that the player can track down the wildcat that caused the damage. Player can then fight the wildcat, try and trap it, or leave. There are two ways to trap the wildcat, one strength based and one dexterity based. If trapped, there are interactions the player can have with the wildcat that they can't by normally fighting them.